added HTTP status to API ProductCategory

+++ src/main/java/com/hyu/electronicsecwebsitebe/controller/ProductCategoryController.java
@@ -4,6 +4,7 @@ package com.hyu.electronicsecwebsitebe.controller;
 import com.hyu.electronicsecwebsitebe.model.ProductCategory;
 import com.hyu.electronicsecwebsitebe.service.impl.ProductCategoryServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,22 +16,35 @@ public class ProductCategoryController {
     private ProductCategoryServiceImpl productCategoryService;
 
     @GetMapping("/all")
-    public List<ProductCategory> getAllProductCategories() {
-        return productCategoryService.getAllProductCategories ();
+    public ResponseEntity<List<ProductCategory>> getAllProductCategories() {
+        List<ProductCategory> listProductCategories = productCategoryService.getAllProductCategories ();
+        return ResponseEntity.ok (listProductCategories);
     }
 
     @GetMapping("/{id}")
-    public ProductCategory getProductCategoryById(@PathVariable String id) {
-        return productCategoryService.findById (id);
+    public ResponseEntity<ProductCategory> getProductCategoryById(@PathVariable String id) {
+        if (!productCategoryService.existsById (id)) {
+            return ResponseEntity.notFound ().build ();
+        }
+        ProductCategory foundProductCategory = productCategoryService.findById (id);
+        return ResponseEntity.ok (foundProductCategory);
     }
 
     @PostMapping("/save")
-    public ProductCategory saveProductCategory(@RequestBody ProductCategory productCategory) {
-        return productCategoryService.saveProductCategory (productCategory);
+    public ResponseEntity<ProductCategory> saveProductCategory(@RequestBody ProductCategory productCategory) {
+        if (productCategoryService.existsById (productCategory.getId ())) {
+            return ResponseEntity.badRequest ().build ();
+        }
+        ProductCategory savedProductCategory = productCategoryService.saveProductCategory (productCategory);
+        return ResponseEntity.status (201).body (savedProductCategory);
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteById(@PathVariable String id) {
+    public ResponseEntity<Void> deleteProductCategory(@PathVariable String id) {
+        if (!productCategoryService.existsById (id)) {
+            return ResponseEntity.notFound ().build ();
+        }
         productCategoryService.deleteById (id);
+        return ResponseEntity.noContent ().build ();
     }
 }
